directory restructure and fixed dependency
