+++ datasetinsights/datasets/dummy/__init__.py
@@ -1,0 +1,2 @@
+# flake8: noqa
+from .dummy_object_detection import DummyDetection2D
