+++ datasetinsights/datasets/protos/string_int_label_map.proto
