+++ datasetinsights/datasets/simulation/__init__.py
@@ -1,6 +1,14 @@
+# flake8: noqa
 from .captures import Captures
+from .download import (
+    Downloader,
+    _filter_unsuccessful_attempts,
+    download_manifest,
+)
+from .exceptions import DefinitionIDError
 from .metrics import Metrics
 from .references import AnnotationDefinitions, Egos, MetricDefinitions, Sensors
+from .tables import SCHEMA_VERSION, FileType, glob
 
 __all__ = [
     AnnotationDefinitions,
