+++ datasetinsights/datasets/synthetic.py
@@ -9,12 +9,13 @@ from sklearn.model_selection import train_test_split
 
 import datasetinsights.constants as const
 from datasetinsights.data.bbox import BBox2D
-from datasetinsights.data.simulation import AnnotationDefinitions, Captures
-from datasetinsights.data.simulation.download import (
+from datasetinsights.datasets.simulation import (
+    SCHEMA_VERSION,
+    AnnotationDefinitions,
+    Captures,
     Downloader,
     download_manifest,
 )
-from datasetinsights.data.simulation.tables import SCHEMA_VERSION
 
 from .base import Dataset
